Move info spot posters to separate component
